feat(reorganizacion back): reorganizado para mayor similitud con MVC
