+++ Project/backend/src/services/contributionsService.ts
@@ -1,6 +1,6 @@
 import simpleGit from "simple-git";
-import { prepareRepo, cleanRepo } from "../../utils/gitRepoUtils"; 
-import { normalizePath, isBinaryFile } from "../../utils/contributions.utils";
+import { prepareRepo, cleanRepo } from "../utils/gitRepoUtils";
+import { normalizePath, isBinaryFile } from "../utils/contributions.utils";
 
 interface ContributionStats {
   [path: string]: {
