+++ Project/backend/src/services/userStatsService.ts
@@ -1,6 +1,6 @@
 import simpleGit from "simple-git";
-import { prepareRepo, cleanRepo } from "../../utils/gitRepoUtils"; 
-import { getPullRequestsByUser, getIssuesByUser, getCommentsByUser } from "../github/githubService";
+import { prepareRepo, cleanRepo } from "../utils/gitRepoUtils"; 
+import { getPullRequestsByUser, getIssuesByUser, getCommentsByUser } from "../services/githubService";
 import { Parser } from "json2csv";
 
 interface UserStats {
